:truck: move scene into separate module
