+++ crates/audioware/src/engine/scene/emitter/mod.rs
@@ -1,0 +1,5 @@
+mod emitter;
+mod emitters;
+
+pub use emitter::Emitter;
+pub use emitters::{Emitters, EMITTERS};
